[Chore] Clean comments, files
