+++ FillMe/FillMe/Scenes/Home/HomeViewController.swift
@@ -8,20 +8,6 @@
 import UIKit
 import Firebase
 
-struct MonthMetadata {
-    let numberOfDays: Int
-    let firstDay: Date
-    let firstDayWeekday: Int
-}
-
-struct Day {
-    let date: Date // represents a given day in a month
-    let number: String // display on the collection view cell
-    let isSelected: Bool // keeps track of whether this date is selected
-    let isWithinDisplayedMonth: Bool // tracks if this date is within the currently viewed month
-}
-
-
 final class HomeViewController: UIViewController {
     
     //MARK:- Views
