+++ FillMe/FillMe/Scenes/Models/DiaryModels.swift
@@ -1,17 +1,11 @@
 //
-//  User.swift
+//  DiaryModels.swift
 //  FillMe
 //
-//  Created by Sue Cho on 2021/01/31.
+//  Created by Sue Cho on 2021/02/24.
 //
 
 import Foundation
-
-struct User {
-    var name: String
-    var email: String
-    var uid: String
-}
 
 struct Diary {
     var date: String
