+++ FillMe/FillMe/Scenes/Models/UserModels.swift
@@ -1,0 +1,14 @@
+//
+//  UserModels.swift
+//  FillMe
+//
+//  Created by Sue Cho on 2021/02/24.
+//
+
+import Foundation
+
+struct User {
+    var name: String
+    var email: String
+    var uid: String
+}
